requirements.txt
Django>=3.2,<4.0
djangorestframework
python-decouple
psycopg2-binary
django-allauth
requests
boto3
django-storages
celery